Changes made to the PR as suggested.
