+++ tests/bootstrap.php
@@ -16,4 +16,4 @@
 
 	require_once __DIR__ . '/../vendor/autoload.php';
 	require_once __DIR__ . '/../vendor/antecedent/patchwork/Patchwork.php';
-	require_once __DIR__ . '/../src/classes/class-spiderblocker.php';
+	require_once __DIR__ . '/../src/index.php';
